Improved database
